# Skip checking block type in sorbet-runtime when possible